requirements.txt
BeautifulSoup==3.2.1
Django==1.7.0
Pillow==2.4.0
bitly-api==0.3
coverage==3.7.1
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.2.1
django-nose==1.2
django-pylibmc==0.5.0
django-suit-redactor==0.0.2
django-reversion==1.8.1
easy-thumbnails==2.0.1
enum34==0.9.23
frosted==1.4.0
humanize==0.5
ipaddress==1.0.6
isort==3.6.2
MySQL-python==1.2.5
natsort==3.1.1
pies==2.6.1
pies2overrides==2.6.1
psycopg2==2.5.3
pylibmc==1.3.0
pytest-django==2.6.2
pytz==2014.4
requests==2.3.0
selenium==2.41.0
the-real-django-wordpress==0.10.1
twitter==1.14.3
wsgiref==0.1.2

-e git://github.com/coleifer/micawber#egg=micawber
-e git://github.com/MotherNatureNetwork/django-suit@titusz-fix-dj17-NoReverseMatch#egg=django-suit
